Implement k-group reversal in q_reverseK
Add support for reversing nodes in groups of k in q_reverseK().
Use extra 2 list_head for store a group of k nodes temporarily,
reverse the temp list_head, then add the nodes in the temp list_head
to final_list.
If the number of nodes is not a multiple of k, the remaining nodes
are left as is.

Change-Id: Ic86792c576fa2454c3704b7fdf2721110620a772