Fix q_delete_dup code defect
Previously, if the last element in the list was a duplicate, it would
not be moved to the pending list due to missing handling logic after
the main duplicate check.

In this patch, a condition statement is added to handle such situation.

This fixes a code defect where some duplicates may remain in the list
unprocessed.

Change-Id: I3bbf9e6522e3979ee7b713d668736ad0254545ab